requirements.txt
fastapi
uvicorn[standard]
python-dotenv
faiss-cpu
sentence-transformers
pypdf
numpy
tiktoken
scikit-learn